+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - [BREAKING] Rename `miden-proving-service-client` to `miden-remote-prover-client` (#1004).
 - [BREAKING] Rename `RemoteProverError` to `RemoteProverClientError` (#1004).
 - [BREAKING] Rename `ProvingServiceError` to `RemoteProverError` (#1004).
+- [BREAKING] Renamed `Note` to `CommittedNote`, and `NetworkNote` to `Note` in the proto messages (#1022).
 
 ### Fixes